+++ crates/edit_prediction/src/capture_example.rs
@@ -1,14 +1,15 @@
 use crate::{
     StoredEvent,
+    data_collection::{compute_uncommitted_diff, uncommitted_diff_for_events},
     example_spec::{ExampleSpec, RecentFile},
 };
 use anyhow::Result;
-use buffer_diff::{BufferDiff, BufferDiffSnapshot};
+use buffer_diff::BufferDiff;
 use collections::HashMap;
-use gpui::{App, AsyncApp, Entity, Task};
+use gpui::{App, Entity, Task};
 use language::Buffer;
-use project::{Project, WorktreeId};
-use std::{collections::hash_map, fmt::Write as _, ops::Range, path::Path, sync::Arc};
+use project::Project;
+use std::{fmt::Write, ops::Range, path::Path, sync::Arc};
 use text::Point;
 
 pub fn capture_example(
@@ -41,7 +42,7 @@ pub fn capture_example(
     let revision = repository_snapshot.head_commit.as_ref()?.sha.to_string();
 
     Some(cx.spawn(async move |cx| {
-        let (uncommitted_diff, events) = uncommitted_diff_for_events(
+        let (uncommitted_diff_events, events) = uncommitted_diff_for_events(
             project.clone(),
             worktree_id,
             root_name.clone(),
@@ -50,6 +51,10 @@ pub fn capture_example(
             cx,
         )
         .await?;
+        let uncommitted_diff = cx
+            .background_executor()
+            .spawn(async move { compute_uncommitted_diff(uncommitted_diff_events) })
+            .await;
 
         let line_comment_prefix = snapshot
             .language()
@@ -122,74 +127,13 @@ pub fn capture_example(
     }))
 }
 
-fn strip_root_name<'a>(path: &'a Path, root_name: &str) -> &'a Path {
-    path.strip_prefix(root_name).unwrap_or(path)
-}
-
-pub(crate) async fn uncommitted_diff_for_events(
-    project: Entity<Project>,
-    worktree_id: WorktreeId,
-    root_name: String,
-    mut events: Vec<StoredEvent>,
-    uncommitted_diffs_by_path: HashMap<Arc<Path>, Entity<BufferDiff>>,
-    cx: &mut AsyncApp,
-) -> Result<(String, Vec<StoredEvent>)> {
-    let mut diff_buffers_by_path: HashMap<Arc<Path>, (Entity<Buffer>, Entity<BufferDiff>)> =
-        HashMap::default();
-    for stored_event in &events {
-        let zeta_prompt::Event::BufferChange { path, .. } = stored_event.event.as_ref();
-        let Some((project_path, relative_path)) = project.read_with(cx, |project, cx| {
-            let project_path = project
-                .find_project_path(path, cx)
-                .filter(|path| path.worktree_id == worktree_id)?;
-            let relative_path: Arc<Path> = project_path.path.as_std_path().into();
-            Some((project_path, relative_path))
-        }) else {
-            continue;
-        };
-
-        if let hash_map::Entry::Vacant(entry) = diff_buffers_by_path.entry(relative_path) {
-            let Some(diff) = uncommitted_diffs_by_path.get(entry.key()).cloned() else {
-                continue;
-            };
-            let buffer = project
-                .update(cx, |project, cx| {
-                    project.open_buffer(project_path.clone(), cx)
-                })
-                .await?;
-            entry.insert((buffer, diff));
-        }
-    }
-
-    events.retain(|stored_event| {
-        let zeta_prompt::Event::BufferChange { path, .. } = stored_event.event.as_ref();
-        let relative_path = strip_root_name(path, &root_name);
-        diff_buffers_by_path.contains_key(relative_path)
-    });
-
-    let uncommitted_diff_snapshots = diff_buffers_by_path
-        .into_iter()
-        .map(|(relative_path, (buffer, diff))| {
-            let snapshot = buffer.update(cx, |buffer, _| buffer.snapshot());
-            let diff_snapshot = diff.update(cx, |diff, cx| diff.snapshot(cx));
-            (relative_path, (snapshot, diff_snapshot))
-        })
-        .collect();
-    let uncommitted_diff = cx
-        .background_executor()
-        .spawn(async move { compute_uncommitted_diff(uncommitted_diff_snapshots) })
-        .await;
-
-    Ok((uncommitted_diff, events))
-}
-
 pub(crate) fn write_event_with_relative_paths(
     output: &mut String,
     event: &zeta_prompt::Event,
     root_name: &str,
 ) {
     fn write_relative_path(output: &mut String, path: &Path, root_name: &str) {
-        for component in strip_root_name(path, root_name).components() {
+        for component in path.strip_prefix(root_name).unwrap_or(path).components() {
             output.push('/');
             write!(output, "{}", component.as_os_str().to_string_lossy()).ok();
         }
@@ -243,107 +187,6 @@ fn compute_cursor_excerpt(
         cursor_in_context,
         context_buffer_start..context_buffer_end,
     )
-}
-
-fn compute_uncommitted_diff(
-    snapshots_by_path: HashMap<Arc<Path>, (language::BufferSnapshot, BufferDiffSnapshot)>,
-) -> String {
-    let mut uncommitted_diff = String::new();
-    let mut snapshots_by_path = snapshots_by_path.into_iter().collect::<Vec<_>>();
-    snapshots_by_path.sort_by(|(left_path, _), (right_path, _)| left_path.cmp(right_path));
-    for (relative_path, (buffer_snapshot, diff_snapshot)) in snapshots_by_path {
-        let base_snapshot = diff_snapshot.base_text();
-        let is_existing_file = diff_snapshot.base_text_exists();
-
-        let new_path_str = relative_path.to_string_lossy();
-        let old_path_str = if is_existing_file {
-            new_path_str.as_ref()
-        } else {
-            "/dev/null"
-        };
-        writeln!(
-            uncommitted_diff,
-            "--- {}{old_path_str}",
-            if is_existing_file { "a/" } else { "" }
-        )
-        .ok();
-        writeln!(uncommitted_diff, "+++ b/{new_path_str}").ok();
-
-        if !is_existing_file {
-            let new_text = buffer_snapshot.text();
-            writeln!(
-                uncommitted_diff,
-                "@@ -0,0 +1,{} @@",
-                new_text.lines().count()
-            )
-            .ok();
-            for line in new_text.lines() {
-                writeln!(uncommitted_diff, "+{line}").ok();
-            }
-            continue;
-        }
-
-        let mut ranges: Vec<(Range<u32>, Range<u32>)> = Vec::new();
-        for hunk in (&diff_snapshot).hunks(&buffer_snapshot) {
-            let old_start = base_snapshot
-                .offset_to_point(hunk.diff_base_byte_range.start)
-                .row;
-            let old_end =
-                exclusive_end_row(base_snapshot.offset_to_point(hunk.diff_base_byte_range.end));
-            let new_start = hunk.range.start.row;
-            let new_end = exclusive_end_row(hunk.range.end);
-            let old_range = old_start.saturating_sub(3)..old_end + 3;
-            let new_range = new_start.saturating_sub(3)..new_end + 3;
-
-            if let Some((last_old_range, last_new_range)) = ranges.last_mut()
-                && (old_range.start <= last_old_range.end || new_range.start <= last_new_range.end)
-            {
-                last_old_range.end = last_old_range.end.max(old_range.end);
-                last_new_range.end = last_new_range.end.max(new_range.end);
-                continue;
-            }
-            ranges.push((old_range, new_range));
-        }
-
-        for (old_range, new_range) in ranges {
-            uncommitted_diff.push_str(&language::unified_diff_with_offsets(
-                &base_snapshot
-                    .text_for_range(
-                        Point::new(old_range.start, 0)
-                            ..row_start_or_max(base_snapshot, old_range.end),
-                    )
-                    .collect::<String>(),
-                &buffer_snapshot
-                    .text_for_range(
-                        Point::new(new_range.start, 0)
-                            ..row_start_or_max(&buffer_snapshot, new_range.end),
-                    )
-                    .collect::<String>(),
-                old_range.start,
-                new_range.start,
-            ));
-        }
-        if !uncommitted_diff.ends_with('\n') {
-            uncommitted_diff.push('\n');
-        }
-    }
-    uncommitted_diff
-}
-
-fn row_start_or_max(snapshot: &language::BufferSnapshot, row: u32) -> Point {
-    if row >= snapshot.max_point().row {
-        snapshot.max_point()
-    } else {
-        Point::new(row, 0)
-    }
-}
-
-fn exclusive_end_row(point: Point) -> u32 {
-    if point.column == 0 {
-        point.row
-    } else {
-        point.row + 1
-    }
 }
 
 fn generate_timestamp_name() -> String {
